Fix empty list indexing in torch._numpy advanced indexing
Issue: Empty list indexing like arr[:, []] failed because torch.tensor([])
creates float32 tensors by default, but PyTorch indexing requires integer tensors.

Fix: Specify dtype=torch.long when converting list indices to tensors.
Added test case to prevent regression.

+++ torch/_numpy/_ndarray.py
@@ -187,7 +187,7 @@ def _analyze_numpy_advanced_indexing(index):
     converted_index = []
     for idx in index:
         if isinstance(idx, list):
-            converted_index.append(torch.tensor(idx))
+            converted_index.append(torch.tensor(idx, dtype=torch.long))
         else:
             converted_index.append(idx)
     converted_index = tuple(converted_index)
